backend/pyproject.toml
[project]
name = "expense-tracker-api"
version = "0.1.0"
description = "Chat-first family expense tracker backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlmodel>=0.0.22",
  "httpx>=0.27.2",
  "asyncpg>=0.30.0",
  "aiosqlite>=0.20.0",
  "pydantic-settings>=2.6.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "pwdlib>=0.3.0",
  "argon2-cffi>=23.1.0",
  "email-validator>=2.2.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
